requirements.txt
djangorestframework==3.14.0
django-environ==0.10.0
requests==2.28.2
pytz==2023.3